Change DlgEditor file name
