+++ src/Gui/DlgSettingsEditor.ui
